requirements.txt
#major
os
sys

# Streamlit for UI
streamlit

# NLP & Text Processing
transformers
torch
sentence-transformers
scikit-learn
nltk
spacy
pandas
numpy

# Topic Modeling
bertopic
faiss-cpu  # For efficient similarity search

# Sentiment & Emotion Analysis
vaderSentiment
textblob

# Miscellaneous
matplotlib
seaborn

#Google cloud access
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
deep_translator